# Improve StringArray(Utf8) sort performance (~2-4x faster)