# Which issue does this PR close?

Improve StringArray(Utf8) sort performance

- Closes [#7847](https://github.com/apache/arrow-rs/issues/7847)

# Rationale for this change
Support prefix compare, and i optimized it to u32 prefix, and u64 increment compare, it will have best performance when experimenting.


# What changes are included in this PR?

Support prefix compare, and i optimized it to u32 prefix, and u64 increment compare, it will have best performance when experimenting.

# Are these changes tested?
Yes


```rust
critcmp  issue_7847  main --filter "sort string\["
group                                         issue_7847                             main
-----                                         ----------                             ----
sort string[0-400] nulls to indices 2^12      1.00     51.4±0.56µs        ? ?/sec    1.19     61.0±1.02µs        ? ?/sec
sort string[0-400] to indices 2^12            1.00     96.5±1.63µs        ? ?/sec    1.23    118.3±0.91µs        ? ?/sec
sort string[10] dict nulls to indices 2^12    1.00     72.4±1.00µs        ? ?/sec    1.00     72.5±0.61µs        ? ?/sec
sort string[10] dict to indices 2^12          1.00    137.1±1.51µs        ? ?/sec    1.01    138.1±1.06µs        ? ?/sec
sort string[10] nulls to indices 2^12         1.00     47.5±0.69µs        ? ?/sec    1.18     56.3±0.56µs        ? ?/sec
sort string[10] to indices 2^12               1.00     86.4±1.37µs        ? ?/sec    1.20    103.5±1.13µs        ? ?/sec
```

# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.