@alamb 2025-07-03
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1016-gcp #16~24.04.1-Ubuntu SMP Wed May 28 02:40:52 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing issue_7847 (0c665fb30903145d071f5aa674a3571af0062836) to 52ad7d703acff0e5b4c143c179df206a842bf24e [diff](https://github.com/apache/arrow-rs/compare/52ad7d703acff0e5b4c143c179df206a842bf24e..0c665fb30903145d071f5aa674a3571af0062836)
BENCH_NAME=sort_kernel
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench sort_kernel 
BENCH_FILTER=
BENCH_BRANCH_NAME=issue_7847
Results will be posted here when complete