@alamb 2025-07-03
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                                   issue_7847                             main
-----                                                   ----------                             ----
lexsort (bool, bool) 2^12                               1.00    116.4±0.34µs        ? ?/sec    1.00    116.5±0.38µs        ? ?/sec
lexsort (bool, bool) nulls 2^12                         1.05    164.4±0.32µs        ? ?/sec    1.00    156.2±0.48µs        ? ?/sec
lexsort (f32, f32) 2^10                                 1.00     45.1±0.14µs        ? ?/sec    1.00     45.2±0.10µs        ? ?/sec
lexsort (f32, f32) 2^12                                 1.00    211.1±2.55µs        ? ?/sec    1.00    212.0±0.40µs        ? ?/sec
lexsort (f32, f32) 2^12 limit 10                        1.00     38.0±0.06µs        ? ?/sec    1.02     38.5±0.07µs        ? ?/sec
lexsort (f32, f32) 2^12 limit 100                       1.01     41.2±0.17µs        ? ?/sec    1.00     40.8±0.20µs        ? ?/sec
lexsort (f32, f32) 2^12 limit 1000                      1.00     78.1±0.13µs        ? ?/sec    1.01     79.1±0.53µs        ? ?/sec
lexsort (f32, f32) 2^12 limit 2^12                      1.00    211.5±0.49µs        ? ?/sec    1.01    212.8±1.18µs        ? ?/sec
lexsort (f32, f32) nulls 2^10                           1.03     54.8±0.14µs        ? ?/sec    1.00     53.4±0.12µs        ? ?/sec
lexsort (f32, f32) nulls 2^12                           1.02    256.1±0.59µs        ? ?/sec    1.00    251.9±0.38µs        ? ?/sec
lexsort (f32, f32) nulls 2^12 limit 10                  1.04     88.7±0.16µs        ? ?/sec    1.00     85.0±0.41µs        ? ?/sec
lexsort (f32, f32) nulls 2^12 limit 100                 1.05     89.8±0.18µs        ? ?/sec    1.00     85.9±0.54µs        ? ?/sec
lexsort (f32, f32) nulls 2^12 limit 1000                1.06    101.5±0.57µs        ? ?/sec    1.00     95.8±0.62µs        ? ?/sec
lexsort (f32, f32) nulls 2^12 limit 2^12                1.02    256.1±0.46µs        ? ?/sec    1.00    252.0±0.45µs        ? ?/sec
rank f32 2^12                                           1.02     69.4±0.16µs        ? ?/sec    1.00     68.2±0.12µs        ? ?/sec
rank f32 nulls 2^12                                     1.01     38.0±0.09µs        ? ?/sec    1.00     37.7±0.11µs        ? ?/sec
rank string[10] 2^12                                    1.00    232.8±0.56µs        ? ?/sec    1.00    233.3±0.56µs        ? ?/sec
rank string[10] nulls 2^12                              1.00    116.7±0.22µs        ? ?/sec    1.00    117.0±0.19µs        ? ?/sec
sort f32 2^12                                           1.00     60.1±0.60µs        ? ?/sec    1.08     65.1±0.78µs        ? ?/sec
sort f32 nulls 2^12                                     1.00     31.8±0.10µs        ? ?/sec    1.01     32.2±0.09µs        ? ?/sec
sort f32 nulls to indices 2^12                          1.00     68.2±0.18µs        ? ?/sec    1.03     70.5±0.27µs        ? ?/sec
sort f32 to indices 2^12                                1.00     72.2±0.15µs        ? ?/sec    1.07     77.3±0.26µs        ? ?/sec
sort i32 2^10                                           1.00      7.5±0.02µs        ? ?/sec    1.14      8.6±0.03µs        ? ?/sec
sort i32 2^12                                           1.00     36.3±0.10µs        ? ?/sec    1.17     42.4±0.12µs        ? ?/sec
sort i32 nulls 2^10                                     1.07      5.8±0.01µs        ? ?/sec    1.00      5.4±0.01µs        ? ?/sec
sort i32 nulls 2^12                                     1.08     24.7±0.07µs        ? ?/sec    1.00     22.9±0.07µs        ? ?/sec
sort i32 nulls to indices 2^10                          1.01     12.5±0.03µs        ? ?/sec    1.00     12.4±0.04µs        ? ?/sec
sort i32 nulls to indices 2^12                          1.00     53.1±0.11µs        ? ?/sec    1.00     53.0±0.09µs        ? ?/sec
sort i32 to indices 2^10                                1.01     11.6±0.02µs        ? ?/sec    1.00     11.5±0.04µs        ? ?/sec
sort i32 to indices 2^12                                1.00     55.1±0.25µs        ? ?/sec    1.00     55.1±0.25µs        ? ?/sec
sort primitive run 2^12                                 1.12      7.1±0.02µs        ? ?/sec    1.00      6.4±0.01µs        ? ?/sec
sort primitive run to indices 2^12                      1.00      8.6±0.02µs        ? ?/sec    1.04      9.0±0.02µs        ? ?/sec
sort string[0-400] nulls to indices 2^12                1.00    182.8±0.59µs        ? ?/sec  
sort string[0-400] to indices 2^12                      1.00    344.4±0.60µs        ? ?/sec  
sort string[10] dict nulls to indices 2^12              1.00    168.9±0.30µs        ? ?/sec    1.01    170.2±0.53µs        ? ?/sec
sort string[10] dict to indices 2^12                    1.00    298.1±0.46µs        ? ?/sec    1.00    299.2±1.39µs        ? ?/sec
sort string[10] nulls to indices 2^12                   1.37    185.9±0.25µs        ? ?/sec    1.00    136.2±0.22µs        ? ?/sec
sort string[10] to indices 2^12                         1.46    336.0±0.50µs        ? ?/sec    1.00    229.8±0.49µs        ? ?/sec
sort string_view[0-400] nulls to indices 2^12           1.02     83.5±0.18µs        ? ?/sec    1.00     82.3±0.20µs        ? ?/sec
sort string_view[0-400] to indices 2^12                 1.00    122.6±0.26µs        ? ?/sec    1.00    122.1±0.24µs        ? ?/sec
sort string_view[10] nulls to indices 2^12              1.02    110.8±0.34µs        ? ?/sec    1.00    109.1±0.25µs        ? ?/sec
sort string_view[10] to indices 2^12                    1.01    174.9±0.27µs        ? ?/sec    1.00    173.1±0.45µs        ? ?/sec
sort string_view_inlined[0-12] nulls to indices 2^12    1.02    105.4±0.66µs        ? ?/sec    1.00    103.7±0.75µs        ? ?/sec
sort string_view_inlined[0-12] to indices 2^12          1.01    165.0±0.22µs        ? ?/sec    1.00    163.1±0.29µs        ? ?/sec
```


</p>
</details>