@Dandandan 2025-07-03
I think this idea is really interesting to compare ! I think we should be able to push this even further a bit by having the prefix computed upfront once (and remove the length) to reduce accessing the string data (so we can mostly do sorting in-place).